Stick footer to the bottom in browsers supporting flexbox
Tested in:
 IE 8/9/10/11
 Safari 5.0
 Chrome
 Firefox
 Opera
Even usable in IE7.
Flexbox is properly supported in IE11/Firefox/Chrome/Opera (nearly 75%
of the browsers right now, according to StatCounter).

In supporting browsers, the footer sticks to the bottom of the screen
with short pages, and behaves as it already does with longer pages.
In all other browsers, the old behavior is kept (minimum height of body
is 300px, and the rest of the page including the footer stays black).